protect delete post

+++ app/routes.py
@@ -181,19 +181,20 @@ def delete_post():
     
     if form.confirm.data:
         post = UserContents.query.filter(UserContents.id == post_id).first()
-        link_contents = LinkContents.query.filter(
+        if current_user.id == post.user_id:
+            link_contents = LinkContents.query.filter(
             LinkContents.content_id == post_id
             ).all()
         
-        delete_file_in_dir(link_contents)
+            delete_file_in_dir(link_contents)
         
-        for link in link_contents:
-            db.session.delete(link)
+            for link in link_contents:
+                db.session.delete(link)
             
-        db.session.delete(post)
-        db.session.commit()
+            db.session.delete(post)
+            db.session.commit()
         
-        return redirect(redirect_url)
+            return redirect(redirect_url)
     
     return redirect(redirect_url)
